Reorganize core compositor module into a folder.
