+++ podpac/compositor.py
@@ -4,4 +4,4 @@ Compositor Public Module
 
 # REMINDER: update api docs (doc/source/user/api.rst) to reflect changes to this file
 
-from podpac.core.compositor import OrderedCompositor
+from podpac.core.compositor.ordered_compositor import OrderedCompositor
